requirements.txt
bs4
asyncurban
spectra
lottie
cairosvg
git+https://github.com/nficano/pytube
pillow
wakeonlan
psutil
ffmpeg
moviepy
redis
dnspython
humanize
pydrive
wikipedia
lxml
search-engine-parser
speedtest-cli
telegraph
telethon
urbandict
pytz
requests
googletrans
gtts
hachoir
pybase64
aiohttp
cowpy
emoji
gTTS-token
gTTS
gitpython
google-api-python-client
oauth2client
google_images_download
SQLAlchemy
psycopg2-binary
spectra
google-cloud-speech
google-cloud-texttospeech
mutagen
python-dotenv
pydownload
pylast
pymongo
Randomwords
lyricsgenius
convertapi
pdf2image
cairocffi
random-password-generator
youtube_search
testresources
spotipy
grpcio
m3u8_To_MP4
osu.py
